# ratings/api/swagger.yaml
basePath: /
definitions:
  models.Rating:
    properties:
      comment:
        type: string
      id:
        type: string
      rating:
        type: integer
      song_id:
        type: string
      user_id:
        type: string
    type: object
info:
  contact:
    email: moran.toulisse@etu.uca.fr
    name: Moran Toulisse.
  description: Spotifaux's renowned API.
  title: Spotifaux
  version: 1.0.0
paths:
  /ratings:
    get:
      description: Get ratings.
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Rating'
            type: array
        "500":
          description: Something went wrong
      summary: Get ratings.
      tags:
      - ratings
    post:
      description: Create a Rating.
      parameters:
      - description: Rating object
        in: body
        name: body
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Rating'
        "500":
          description: Something went wrong
      summary: Create a Rating.
      tags:
      - ratings
  /ratings/{id}:
    delete:
      description: Delete a Rating.
      parameters:
      - description: Rating UUID formatted ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "422":
          description: Cannot parse id
        "500":
          description: Something went wrong
      summary: Delete a Rating.
      tags:
      - ratings
    get:
      description: Get a rating.
      parameters:
      - description: Rating UUID formatted ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Rating'
        "422":
          description: Cannot parse id
        "500":
          description: Something went wrong
      summary: Get a rating.
      tags:
      - ratings
    put:
      description: Update a Rating.
      parameters:
      - description: Rating UUID formatted ID
        in: path
        name: id
        required: true
        type: string
      - description: Rating object
        in: body
        name: body
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "422":
          description: Cannot parse id
        "500":
          description: Something went wrong
      summary: Update a Rating.
      tags:
      - ratings
schemes:
- http
swagger: "2.0"